ux: Rename `import-url` API route to `import`
